+++ firebase-ai/src/test/java/com/google/firebase/ai/SerializationTests.kt
@@ -16,6 +16,8 @@
 
 package com.google.firebase.ai
 
+import com.google.firebase.ai.common.TemplateGenerateContentRequest
+import com.google.firebase.ai.common.TemplateGenerateImageRequest
 import com.google.firebase.ai.common.util.descriptorToJson
 import com.google.firebase.ai.type.Candidate
 import com.google.firebase.ai.type.CountTokensResponse
@@ -514,6 +516,56 @@ internal class SerializationTests {
       """
         .trimIndent()
     val actualJson = descriptorToJson(Tool.Internal.serializer().descriptor)
+    expectedJsonAsString shouldEqualJson actualJson.toString()
+  }
+
+  @Test
+  fun `test template request serialization as Json`() {
+    val expectedJsonAsString =
+      """
+        {
+          "id": "TemplateGenerateContentRequest",
+          "type": "object",
+          "properties": {
+            "inputs": {
+              "type": "object",
+              "additionalProperties": {
+                "${"$"}ref": "JsonElement"
+              }
+            },
+            "history": {
+              "type": "array",
+              "items": {
+                "${"$"}ref": "Content"
+              }
+            }
+          }
+        }
+      """
+        .trimIndent()
+    val actualJson = descriptorToJson(TemplateGenerateContentRequest.serializer().descriptor)
+    expectedJsonAsString shouldEqualJson actualJson.toString()
+  }
+
+  @Test
+  fun `test template imagen request serialization as Json`() {
+    val expectedJsonAsString =
+      """
+        {
+          "id": "TemplateGenerateImageRequest",
+          "type": "object",
+          "properties": {
+            "inputs": {
+              "type": "object",
+              "additionalProperties": {
+                "${"$"}ref": "JsonElement"
+              }
+            }
+          }
+        }
+      """
+        .trimIndent()
+    val actualJson = descriptorToJson(TemplateGenerateImageRequest.serializer().descriptor)
     expectedJsonAsString shouldEqualJson actualJson.toString()
   }