+++ appcheck/firebase-appcheck-debug-testing/CHANGELOG.md
@@ -80,3 +80,4 @@
 # 16.0.0-beta01
 
 - [feature] Initial beta release of the [app_check] Debug Testing SDK with abuse reduction features.
+