+++ firebase-ai/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Unreleased
 
+# 17.7.0
+
+- [changed] Added `LiveAudioConversationConfig` to control different aspects of the conversation
+  while using the `startAudioConversation` function.
+- [fixed] Fixed an issue causing streaming chat interactions to drop thought signatures. (#7562)
+
+# 17.6.0
+
+- [feature] Added support for server templates via `TemplateGenerativeModel` and
+  `TemplateImagenModel`. (#7503)
+
+# 17.5.0
+
 - [changed] Added better scheduling and louder output for Live API.
 - [changed] Added support for input and output transcription. (#7482)
 - [feature] Added support for sending realtime audio and video in a `LiveSession`.